Broken Access Control Fixed

+++ src/main/java/es/grupo18/jobmatcher/controller/web/BlogController.java
@@ -14,7 +14,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
@@ -221,6 +220,10 @@ public class BlogController {
     @PostMapping("/blog/posts/{postId}/delete")
     public String deletePost(@PathVariable long postId) {
         try {
+            long userId = userService.getLoggedUser().id();
+            if (!postService.canEditOrDeletePost(postId, userId)) {
+                return "error/403";
+            }
             postService.delete(postService.findById(postId));
             return "redirect:/blog/posts";
         } catch (NoSuchElementException e) {
@@ -270,6 +273,10 @@ public class BlogController {
             @RequestParam String text,
             @RequestParam int rating) {
         try {
+            long userId = userService.getLoggedUser().id();
+            if (!reviewService.canEditOrDeleteReview(reviewId, userId)) {
+                return "error/403";
+            }
             if (text.isBlank()) {
                 model.addAttribute("post", postService.findById(postId));
                 model.addAttribute("review", reviewService.findById(reviewId));
@@ -287,6 +294,10 @@ public class BlogController {
     @PostMapping("/blog/posts/{postId}/reviews/{reviewId}/delete")
     public String deleteReview(@PathVariable long postId, @PathVariable long reviewId) {
         try {
+            long userId = userService.getLoggedUser().id();
+            if (!reviewService.canEditOrDeleteReview(reviewId, userId)) {
+                return "error/403";
+            }
             reviewService.delete(reviewId, postService.findById(postId));
             return "redirect:/blog/posts/" + postId;
         } catch (NoSuchElementException e) {
@@ -303,13 +314,13 @@ public class BlogController {
             ReviewDTO review = reviewService.findById(reviewId);
             model.addAttribute("review", review);
 
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
             boolean canEdit = false;
-            if (auth != null
-                    && auth.isAuthenticated()
-                    && !(auth instanceof AnonymousAuthenticationToken)) {
-                String username = auth.getName();
-                canEdit = reviewService.canEditOrDeleteReview(reviewId, username);
+            if (SecurityContextHolder.getContext().getAuthentication() != null &&
+                    SecurityContextHolder.getContext().getAuthentication().isAuthenticated() &&
+                    !(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)) {
+
+                long userId = userService.getLoggedUser().id();
+                canEdit = reviewService.canEditOrDeleteReview(reviewId, userId);
             }
             model.addAttribute("canEditReview", canEdit);
